+++ ClassLibrary/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MSTest.TestFramework" version="2.1.2" targetFramework="net472" />
+</packages>
